Set WebSocket.binaryType to "arraybuffer".
By default if binary data is received on a WebSocket it is handed to the
client script as a DOM Blob object, which must be read asynchronously.  If
this option is set on the WebSocket then binary data is handed to the script
as ArrayBuffer objects, which can easily be read and handled on the fly.